+++ tests/parser/test_address_parser_test_api.py
@@ -1,6 +1,5 @@
 # Since we use a patch as model mock we skip the unused argument error
 # pylint: disable=W0613, too-many-arguments
-import os
 import unittest
 from unittest.mock import patch, call
 
@@ -38,23 +37,6 @@ class AddressParserRetrainTest(AddressParserPredictTestCase):
         cls.a_best_checkpoint = "best"
 
         cls.verbose = False
-
-    def setUp(self):
-        super().setUp()
-
-        self.a_loss_function = nll_loss
-        self.a_list_of_batch_metrics = [accuracy]
-
-        # to create the dirs for dumping the prediction tags since we mock Poutyne that usually will do it
-        os.makedirs(self.a_logging_path, exist_ok=True)
-
-    def tearDown(self) -> None:
-        # cleanup after the tests
-        path = os.path.join(self.a_logging_path, "./prediction_tags.p")
-        if os.path.exists(path):
-            os.remove(path)
-
-        os.rmdir(self.a_logging_path)
 
     def address_parser_test_call(self):
         self.address_parser.test(self.mocked_data_container,
